Utils: Remove pixi module and move find_pixi to conda module
